Fix framework issues
