+++ frameworks/Java/hexagon/.gradle/2.13/taskArtifacts/cache.properties
@@ -1,0 +1,1 @@
+#Mon May 09 00:41:48 CEST 2016
