--- frameworks/Java/hexagon/gradle/wrapper/gradle-wrapper.properties
@@ -1,6 +1,0 @@
-#Sun May 08 23:51:57 CEST 2016
-distributionBase=GRADLE_USER_HOME
-distributionPath=wrapper/dists
-zipStoreBase=GRADLE_USER_HOME
-zipStorePath=wrapper/dists
-distributionUrl=https\://services.gradle.org/distributions/gradle-2.13-bin.zip
