+++ frameworks/Kotlin/hexagon/src/main/resources/benchmark.properties
@@ -1,0 +1,6 @@
+web.port=5050
+web.host=0.0.0.0
+
+mongodb.database=hello_world
+mongodb.world.collection=world
+mongodb.fortune.collection=fortune
