+++ frameworks/Kotlin/hexagon/src/main/resources/logback.xml
@@ -1,0 +1,19 @@
+<!--
+ | Logback configuration for tests
+ |
+ | All output is ignored by default, to enable logs change root logger level.
+ !-->
+<configuration>
+  <appender name="console" class="ch.qos.logback.core.ConsoleAppender">
+    <encoder>
+      <Pattern>
+        %d{HH:mm:ss.SSS} %highlight(%-5p) %magenta([%-10thread]) %-30logger{0} | %m%n
+      </Pattern>
+    </encoder>
+  </appender>
+
+  <root>
+    <level value="off" />
+    <appender-ref ref="console" />
+  </root>
+</configuration>
