+++ frameworks/Kotlin/hexagon/src/main/resources/templates/fortunes.html
@@ -1,0 +1,21 @@
+<!DOCTYPE html>
+
+<html>
+<head>
+  <title>Fortunes</title>
+</head>
+<body>
+<table>
+  <tr>
+    <th>id</th>
+    <th>message</th>
+  </tr>
+  {% for fortune in fortunes %}
+  <tr>
+    <td>{{ fortune.id }}</td>
+    <td>{{ fortune.message }}</td>
+  </tr>
+  {% endfor %}
+</table>
+</body>
+</html>
